+++ Renderer.h
@@ -7,10 +7,10 @@
 
 void draw(const VertexBuffer &vertexbuffer, const IndexBuffer &indexbuffer, const Shader &shader);
 
-void drawRect(const GLfloat position[], const GLfloat color[], const GLuint width,
-              const GLuint height);
+void drawRect(const GLfloat position[], const GLfloat color[], const GLfloat width,
+              const GLfloat height);
 
-void drawRect(const GLfloat position[], const GLuint width, const GLuint height);
+void drawRect(const GLfloat position[], const GLfloat width, const GLfloat height);
 
-void drawRect(const GLfloat x, const GLfloat y, const GLuint width, const GLuint height);
+void drawRect(const GLfloat x, const GLfloat y, const GLfloat width, const GLfloat height);
 #endif
